♻️ refactor : 발견 - 주식종목만 최근검색어로 저장하도록변경
Message body
-해더의 navList.ts 발견,마이페이지 url 수정
