+++ src/app/(afterLogin)/search/[keyword]/page.tsx
