+++ src/common/Header/navList.ts
@@ -1,8 +1,8 @@
 const navList = [
-  { path: "/", name: "발견" },
+  { path: "/search", name: "발견" },
   { path: "/", name: "뉴스" },
   { path: "/", name: "관심종목" },
-  { path: "/", name: "마이페이지" },
+  { path: "/mypage/profile", name: "마이페이지" },
 ];
 
 export default navList;
